Move bounded distributions to new file
